NeXT needs #define pid_t int in includes.h
namedbname.c:find_name_search() needed search capability on both local
subnet records and WINS records. it used to return NULL if the name
was not found in the local subnet records even if you asked it to search
the WINS name records as well.
(This used to be commit 96c2efded884ce4390c5da68afaccca9acc17558)