added code that checks whether the DOMAIN<1b> name is claimed on the local subnet _before_ going and registering it.  the reason for this is that i don't want to cause chaos by registering the name twice.
it's ok to register DOMAIN<1b> with the WINS server, because the WINS
server's job is to check a) _if_ there's a current owner of the name
b) if the current owner exists c) if the current owner still wants the
name.  if they don't, the WINS server responds by saying, 'yes, you can
have it'.

lkcl