connecting to IPC$ goes under the guest account.  the IPC$ share should be treated no differently than any other share (for any security setting: user, server or share).
this will clear up a bug where, when clients connect to the IPC$ share, this
used to be done under the guest account.  the standard_sub_basic() macros
will substitute the _guest_ account for %U, causing the samba server to
look different from when the client then connects to any other share.

lkcl
(This used to be commit 38526569608f6eb5f098efee7013037873d140bf)