This commit fixes the security hole due to buffer overflows.
The main fix is just a couple of lines long, but I have now also put
precautionary checks on a large number of other places in the code
where unchecked string copies were being performed. An exploit via one
of these copies is unlikely, but is is better to be safe.

I also added a routine to log possible exploit attempts using the code
that was posted for obtaining root access on a Samba server.