add become_root()/unbecome_root() calls. These should be used around regions of the code that need root privilages but which normally don't have them (ie. when processing most SMBs). These functions can optionally save/restore the current working directory as well
updated several places in the code that previously used unbecome_user()
to achieve the same thing to now use become_root() and unbecome_root()

changed close_file() to take an additional argument which says whether
this is a normal SMBclose type call or a close that resulted from some
other action, such as the close of a connection. If it is a abnormal
close then don't use magic scripts or print files.

changed sys_utime() to ignore any attempt to set the modification time
to 0 or -1. We've had reports that files have had their time set to
0 so this should catch those.