change this in 1.9.17 as well:
---------
don't do the 255.255.255.255 mapping for group names if the name type
is 0x1c. This allows remote logon servers when using a Samba WINS server.

I did some experiments with a real NT WINS server and this behaviour
is much closer to being correct. It will only be totally correct when
we add proper group name support.