many systems don't have /etc/shadow but do have another system for making encrypted passwords secret. For example, with secure NIS+ only root can get the encrypted password.
hash in the encrypted password of "root" to provide a nice source of
secret on such systems. On systems that don't have this (ie. any user
can get roots encrypted password) then the security is so slack that
it probably doesn't matter what we do to generate the secret.
(This used to be commit 3271e4c29fdc15a1ae61dec94517e484c2457411)