Jeremy is going to hate me :-)
A bug in slprintf() means that although p7 compiled fine under SunOS4
it won't actually run. The problem is that vsprintf returns a pointer
under SunOS not an integer (giving the number of chars printed). I
wonder how many other platforms this affects?

I've fixed it (trivial fix) but now we need to decide when a new
release is needed. Right now SunOS users will be wondering why it is
broken.