I've replaced the linked list used to manage the subnet namelists with a splay tree.  For short lists, this will have no noticable effect.  As lists (eg. the WINS database) grow longer, the speed improvements should be quite dramatic.  This change is an incremental step toward replacing the in-memory namelists with a back-end database.
This change is going into the 1.9.19pre-alpha code because...well...it's
pre-alpha.  Please let me know if there are any problems.

(Oh, as a side-effect, the wins.dat will be in sorted order.  :)

Chris -)-----
(This used to be commit 7806c453df02a89f67e7c5c8b91f24aa2274e756)