The function add_name_to_subnet(), in file nmbd_namelistdb.c, returns a pointer to the newly constructed name list entry.  In most cases, this return value is ignored.  The two exceptions are in asyncdns.c and nmbd_winsproxy.c.
Most of the calls which ignored the return value were not cast to void, so
I added the cast.  This helped me sort out which calls really did use the
return value.  I also discovered one case, in nmbd_winsserver.c, in which
the return value was being stored to a variable which, in turn, was not
used.

Chris -)-----