In addition to the HAVE_FUNCTION_MACRO test that Andrew used as an example, I also need to know if the __FILE__ macro exists.  I followed the example, but ran into two problems:
1) I don't have the autoconf tools installed.
2) The instructions did not say which files should be checked in after
   running autoheader and autoconf.

So I'm checking in the modified acconfig.h and configure.in on the
assumption that the next time someone runs autoheader and autoconf my
new tests will be included.

Hope that's not a problem (and I'll try to grab a copy of autoconf from
somewhere).

Chris -)-----
(This used to be commit 0637a49d5d6c03f10d462be4e92c4e4bbf7ff6c2)