check whether system type is the same as stored in the cache (full stand-alone macro, candidate for inclusion in the next release of autoconf)
create configure flag --enable-maintainer-mode, that enables automatic
rebuilding of configure from configure.in and aclocal.m4, and of
config.h.in based on acconfig.h.  This assumes autoconf and autoheader
are available.  Maintainer mode also introduces automatic dependency
tracking of include files, but it depends on features provided by gcc
and GNU make, so you must not enable maintainer mode if you are not
using any of these
(This used to be commit 64cba1cc8e689bcde2dc124c511da086c33fa4a8)