changed the default filesystem type to NTFS (from Samba)
I've finally changed this because I've found a real effect that this
has. Win95 shows long filenames in a directory listing in a dos box
only if this is set to NTFS or HPFS (only ones I've found so far
anyway).

If NTFS poses problems of some kind then we should try HPFS

I wonder if this change will trigger any attempts to use ACLs ?