drat.
on Linux 2.0 libc detects the lack of getdents in the kernel and used
opendir() etc. so we need to implement those ...

it would be needed for a port to others OSes anyway I suppose.
(This used to be commit e3fd861590dd27cc643c2d8d0cb5a5651c84d9ac)