Try to make smbprint a better behaved citizen.
Removed a -x on the first line that causes lots of output to the console
and added a command line parameter handling flag for the tranlate case.

The problem is that smbprint is being overwritten/replaced when you upgrade
a RedHat rpm.  If you had modified smbprint to include translation, you
loose it, and you have to go and repatch etc.

Now, we have a -t flag on smbprint, so we should have fewer problems in
this regard.

Untested as yet, but soon will be ...